Check if Process non-nil before calling Kill
A bad interleaving could (and has - golang/dep#391) cause the context
cancellation select branch to run before the normal process termination
branch.

+++ cmd.go
@@ -63,8 +63,10 @@ func (c *monitoredCmd) run(ctx context.Context) error {
 				return &timeoutError{c.timeout}
 			}
 		case <-ctx.Done():
-			if err := c.cmd.Process.Kill(); err != nil {
-				return &killCmdError{err}
+			if c.cmd.Process != nil {
+				if err := c.cmd.Process.Kill(); err != nil {
+					return &killCmdError{err}
+				}
 			}
 			return c.ctx.Err()
 		case err := <-done:
